Enabled Multi-fit To Utilise Any IV Fitter
Added a kwarg to IVData.multi_fit() to allow the specification of a
specfic fitter object for use in the 2nd and 3rd stage of the multi-fit
method. Currently the first stage has a fixed fitter, this will soon
also be changeable.

Other files included use of this new functionality in the analysis
script and an additional doc string for the 3-param fitter.