Edit IV Comparison Plot to use Vf Approximation
The IV comparison plot changed to use an alternatively calculated V_f,
instead of the fitted value of V_f used in the calculation of the
analytical TS IV curve, the approximation (taken from Stangeby) of 3 T_e
is used instead. Legend labels are also updated to list the density and
temperatures of each IV curve.