Alter Spice Analysis for Changes in Hist Finding
Alterations are made to the functions run_multi_hist_analysis() and
run_histogram_extr() to reflect the prior changes made in the default
behaviour of the histogram region finding - notably the
run_histogram_extr() function has been deprecated as it was basically
ported to be an official part of splopter.

This commit is part of the wider changes to the functionality of the
histogram binning region.