Restructure into Package with Core Subpackage
Flopter is reorganised into a more package-like structure, with the root
containing docs, data, and the flopter package proper. Within flopter,
the core subpackage is added to the previously existing subpackages
(analysis, classes, databases, external, functions, scripts) and the
two utils modules are moved to a new utils package. Sphinx index and
config files are updated to reflect this change.

NOTE: This change is not final, more restructuring into fewer
subpackages is likely before too long.