Fix and Clean-up Documentaton Generation
The automatic documentation generation using Sphinx and Autodoc is fixed
following the refactor of core flopter modules into a different
structure. The gitlab continuous integration is used to build the html
source upon pushing changes to the repo, so documentation is
continuously updated - this is specified in the .gitlab-ci.yml.

Additionally, as the documentation is auto-built on the server, the html
does not need to be stored in the repository, so the files are removed
and the folders added to the .gitignore.