Add Offline Magnum LP Analysis Scripts
Large update which adds the capability to run offline (i.e. no
connection to the magnum database required) analysis of Langmuir probe
data taken on the second visit to Magnum.

This involves using a modified magopter (magoptoffline) which runs the
same code, but either ignore Magnum database data or uses a netCDF file
containing all the necessary data instead. This netCDF file is generated
by the database_retrieval.ipynb notebook in examples/magnum-analysis.

Magopter and magoptoffline will need ot be remerged into a single
entity at some point, as this was a quick hack created in the control
room to try and get shot-for-shot analysis running.