Fix Magoptoffline Processing and Initial Analysis
Updated magoptoffline to scale voltage properly through the
pre-acquisition divider based on which shunt voltage is supplied. Also
fixed a flaw whereby raw current was being used to alter the voltage
instead of the actual current.

Initial analysis is updated to be more robust to mismatched numbers of
sweeps. Also forced no downsampling and fixed the get_sweep_range
function.