Update Notebooks for Initial Magnum Data Analysis
The database_retrieval.ipynb notebook is for creating a database of
magnum information using xarray and saving it as a netCDF file. There is
also some boxes for how to feed this information into analysis routines.

The result_concatenation.ipynb notebook is for combining together the
initial analysis output files and doing some initial analysis. So far
initial analysis is restricted to the axial profile scan and there are
currently issues with spurious Nans throughout the arrays.