Update Examples for Concatenating Magnum Analysis
The datasets outputted by the initial run-throughs of Magnum adc
analysis can be stitched together to form a single dataset. These two
jupyter notebooks outline 2 different methods for merging these datasets
together. The second one (result_concatenation_1.ipynb) is better, as it
includes a fix for the issue with the time dimension not aligning
properly and creates a reference dataframe which allows one to more
easily see which files can/should be combined.

This should be merged into the main analysis routines.