Add Protoype Script for Copying MAST-U LPMUX Data
This is the script I created in Eindhoven for running on Alex Poelman's
python integration with the Magnum triggering system. It has 3 different
methods for copy data from the MAST-U multiplexers (LPMUX) via http
requests, depending on what packages were available when running the
script on a Magnum control room computer.