Update MagnumProbes Class for Machined Geometries
Updated the MagnumProbes class to have the corrected geometry for the
DSF probes having been machined for the second experiment at Magnum.
Also changed the labelling to reflect the new standard (S, L, B, R) as
opposed to the old designations (S, R, L, C).

Also got rid of the MagnumProbes.probes dictionary and replaced it with
an override of the __getattr__() in built, so there are no longer two
copies of the probes stored in the class.