Overhaul Package Structure
The structure for the package, and whole repository in fact, is
overhauled to be more concise and sensible. Biggest changes come from
the splitting of previously core packages (splopter, magopter) into
their requisite speciality folders (spice and magnum respectively). The
other folders were then redistributed into these three folders depending
on their specificity. Homogeniser was split into separate declaration
and instance files.

The analysis and scripts folders have been moved out of the flopter
package and into a top-level analysis folder along with the jupyter
notebooks. This is to keep all of the tools in the package and their
uses in analysis.