Refactor and Rename for Consistency
This is an extension of the previous commit, further refactoring and
renaming is done to make the codebase more consistent and readable.
magnumdbutils and magnumdb are merged, normalise and homogenise are
split into core and spice parts, and all submodule files are renamed to
have a less busy import statement.