Add Function For Generating Input File Parameters
A function, generate_2d_input, is added which allows input file
parameters to be printed/plotted from various input parameters
including a LangmuirProbe object, temperature, density, and magnetic
field strength and angle. This also required generalising some of the
outputs of the LangmuirProbe class, such as whether it has an angled tip
or not.