Add xarray Conversion Methods and Update Prepare
The prepare method is updated to correct data appropriately using the
newly calculated offsets on the adc data. The prepare method is also
modified to prevent unnecessary computation (partially) by only
performing peak finding on a single voltage trace.

The method to_xarray is added, which allows for a prepared magopter
object to be converted into a dataset (adapted from the code in the
magnum initial postrun analysis).