Add Analysis Script for Combining All Magnum-2 Data
A python analysis script for combining all the metadata and averaged IV
characteristics of the magnum-2 data into a single dataset which can be
lazily loaded to speed up analysis - instead of having to combine at
analysis time which can take 10 minutes for each set of ~22 datasets.

This is a first draft of this script but based on the routines developed
in the jupyter notebook result_concatenation_*.ipynb.