Update of All Jupyter Notebooks - 2019-11-02
As part of the backup process before switching to Manjaro, all
repositories are being backed up including untracked changes in
notebooks.

This includes the beginnings of reorganisation to more specific spice
and magnum notebooks - as they were getting messy - and improved
multi-splopter analysis for the lowdensity angle scan simulations and
restart tests.