Allow Reduced Tfile Reading and Fix t-zeroing
A reduced dataset can now be requested from a tfile with the kwarg
variable_names. This has to be a list of possible variables in the file.
The member variables not loaded are then filled with None.

T-zeroing upon simulation restart can now be detected and 'accounted'
for by creating a dummy t-array with arange and multiplying by dt.