Refactor Homogenise and Auto Concat Restart Vars
Homogenise is eschewed in favour of an additional method within
IVAnalyser for homogenisation. The Spice-2 homogenise implementation is
placed back into the main Splopter module and updated to allow automatic
concatenation of restarted simulation variables. This is the first
attempt at this and doesn't work entirely consistently - so will be
replaced.

Implementation for autoconcat is based on preliminary work done in
restarttest notebooks.