Update of Analysis Notebooks - 2019-12-23
This is an update of all the analysis jupyter notebooks made for
analysing simulations and magnum data. The simulation analysis consists
of lowdens_anglescan compilation, restart_test concatenation &
comparison, and plotting of simulation input files; the magnum analysis
of refinement of analysis through to analysis of particular shots for
Joe Allcock, as well as compilation into one large dataset and analysis
of that large dataset.

The magnum_postrun_initial_analysis.py is also updated and refactored
for easier testing.