Split prepare() Into Separate Functions
The prepare method in magoptoffline is refactored to be more in line
with the recent changes made to the ivanalyser base class. The
preparation is deemed as the applying of calibration settings to the raw
data and applying a filter. The homogenisation is deemed as the
splitting of the voltage trace into individual sweeps and then
adding them to a numpy array.