Fix Homogenisation and Filtering for Adv. Analysis
The homogenise function, in particular the sweep partitioning section,
has been fixed to be more robust and to work faster. Previously I tried
to speed it up by only iterating once but this has been undone and the
code profiled to figure out where the bottlenecks were, with the calls
to argrelextrema being the major bottlenecks. Taking a smaller slice of
5% of the raw data seems to produce the exact same result and speed up
by orders of magnitude. Robustness was achieved through using the raw
voltage data to partition sweeps as opposed to the calibrated voltage
data, which could have peaks in unexpected places.

The Filtering was also broken as the butterworth filter was only being
applied to the voltage.

NOTE: The sweep partitioning still relies on a questionable where
statement to function properly, this may need to be replaced if problems
with the sweep partitioning remain.