Fix Downsampling and Add Logging Messages
The downsampling, controlled by the desired_datarate variable, was
previously set too high and was therefore not downsampling at all.
Lowered it again to a value of 10000.

Additionally added a logging/print function which appends a datetime
string and the current shot number to a printed message, making it
easier to keep track of where each shot has gotten to in the processing
stage.