Add Option to Minimise Temperature in Multifit
The multi_fit method in IVData gets a major upgrade to be more robust
by having an option to apply the final fit multiple times at different
'trim' lengths and minimise the temperature fitted. This (a) lowers the
temperature fitted by the probes, (b) raises the chi^2 to ~1, and (c)
makes the fit more robust to different orders of magnitude of current.

This involved a general upgrade to IVData to make it more compatible
with xarray, adding methods for inherently getting the floating
potential. Also upgraded the plot() function in fitdata, added some
xarray constants and fixed a problem whereby the FullIVFitter would
potentially get stuck on a bad fit with large sheath expansion parameter
by lowering the upper fit limit on that fit variable.