Add Auto-Concatenation of Backup Simulation Files
The work done on preparing a complete current and voltage profile from
several simulation output files (from a multiply restarted simulation)
is implemented in the main splopter functions, optionally controlled by
a kwarg to the prepare() method. Also now checks automatically for the
percentage completion of the simulation and only continues if it is
sufficiently complete.