Remove Postamble From Input Factory
The postramble, i.e. the code segment under if __name__ == '__main__':
is moved to a separate analysis script to tidy up inputfactory.

The exisiting plotting routines are also slightly updated and an
additional plotting routine, for plotting the angular dependence of a
given probe in an input file in terms of r_L and lambda_D, is also
added.