Add Several Fitters and Guess Generation
The fitter baseclass is updated to include 6 additional fitters, using
the briefly considered experimental models which include the effect of
electron capture through sheath expansion. Also alternative fitters
which calculate a based on density and temperature as opposed to c_1 and
c_2.

Also added the functionality to generate a guess for any given fitter
from either kwargs or default values.