Implement Different Homogenisation Methods
The method by which splopter extracts the IVData objects from the
simulation data is expanded, with a choice of 2 methods now possible.
The first method is the original prepend-downsample used when the V and
I arrays are different lengths. The second method is the new method
devised for spice verison 2.14 and above where in both the V and I
arrays are the same length. This allows for pandas to be used to bin the
current values by voltage, thus giving a better estimate of uncertainty
on each current value in the corresponding IV.

This is controlled by the 'version' kwarg in Splopter's __init__(),
defaulting to version 2.13. The option to store the generated pandas
dataframe as an attribute is also included as the boolean flag kwarg
store_dataframe_fl.

Additional kwargs controlling ignoring behaviour are also added,
including ignore_tzero and ignore_a_file boolean flags. Finally a
function to parse the input file (using InputParser) without invoking
prepare() is added.