Add Method to Estimate Plasma Potential (+others)
A method is added which can estimate the plamsa potential from teh
position of the knee in the electron region. This can be done either by
differeniating the IV characteristic or by fitting a spline.

Additional updates include: an additional parameter to minimise in
fit_to_minimum(), bugfix whereby malformed sigma data could be passed,
updating the xarray conversion method to support electron and ion
current, adding the option to specify a stage 2 guess in multi_fit(),
adding the option to trim to the floating potential in multi_fit(), and
additional minor changes to the plotting in fit_to_minimum().