Add Functions for Alternative Sheath Expansion Params
The alternative sheath expansion parameters derived for chapter 5 are
given explicit functionality within lputils. New functions are created
for the parallelogrammatic and rotated rectangular sheath models and the
decomposition functions also implemented. Methods are added to the probe
base class to both calculate the sheath expansion parameter given probe
geometry, allowing the form of the SE parameter to be specified.

Additionally, several key quantities are added as functions (xi, debye
length, gyrofrequency, sounds speed, larmor radius) as well as some of
their error propagation equations. Z and m are also added into
calculations of density and ion saturation current.

A method for esitmating temperature is also added which takes the plasma
potential and floating potential as arguments. This is rearranged from
the equation for the floating potential.