Fix Arc Filtering and Add Multi-Dataset Functions
This commit is part of a big update combining together recent changes to
core splopter modules.

The first of two main updates in this commit are the fixing of the autosel
functions to properly auto-detect arcs and exclude them from the
selection part of the initial concatenation stages. This is implemented
in the function find_allowed_sweeps() which operated similarly to
find_sweep_lims() but instead returns a list of allowed sweeps rather
than an upper and lower bound. Additional functions for autosel are also
added such as preprocess_average_coarsen which allows for easy
downsampling. Related to this is the new function to plot the sweep
trimming so as to get a visualisation of what the preprocess function is
doing for each shot.

The second major part of the update is the inclusion of several
jupyter-made functions for combining together datasets, namely
fit_multi_magnum_ds and combine_fit_ds which were used in generating the
datasets required for plotting figures in chapter 6 of the thesis.
Interpolating the TS position is also added as a function, but made more
general than the original function. Two plotting functions for the
density scan are also added.

Finally there are a series of changes to the main fit function - a major
part of the magnum analysis workflow - which incorporate recent small
changes to core splopter functionality. These include: adding
phi-estimated temperature values to the outputed fit_df, bugfixing the
density calculations sometimes being returned as a list, making the use
of multi_fit optional, adding the mass and charge (Z) of the particles
to the density calculations, generation of an initial guess for fitting
purposes, adding controllable print output, and passing of kwargs through
to the fitting function.