Add Functions for Analysis of SPICE Simulations
The main functions for use in the analysis of SPICE parameter scans are
added to splopter. This includes the main data process functions
(xrfuncs) and the decomposition functions.

xrfuncs constitutes a workflow for iteratively generating an xarray
dataset, with dimensions corresponding to scanned parameters, containing
all IV data and parameters from a selection of fits. This is adapted
from routines created in jupyter notebooks. It is commited in its final
state as of thesis submission.

The decomposition module contains several functions for decomposing
sheath expansion parameters and plotting them simply. Again this is
adapted from jupyter-made functions.